refactor: break out hook definitions (WIP)
Also, creating new interfaces for DragSource, DropTarget specification interfaces
