+++ packages/react-dnd/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useDragSource'
 export * from './useDropTarget'
 export * from './useMonitorOutput'
+export * from './useDragLayerOutput'
