+++ packages/react-dnd/src/hooks/useDropTarget.ts
@@ -2,19 +2,15 @@ import * as React from 'react'
 import { TargetType } from 'dnd-core'
 import registerTarget from '../registerTarget'
 import createTargetMonitor from '../createTargetMonitor'
-import { DropTargetMonitor, DropTargetSpec } from '../interfaces'
-import {
-	Ref,
-	NoArgs,
-	HandlerManager,
-	useDragDropManager,
-	useMonitorSubscription,
-} from './util'
+import { DropTargetMonitor, DropTargetHookSpec } from '../interfaces'
+import { useDragDropManager } from './useDragDropManager'
+import { useMonitorSubscription } from './useMonitorSubscription'
+import { Ref, HandlerManager } from './util'
 
-function useDropTargetHandler(targetSpec: DropTargetSpec<void>) {
+function useDropTargetHandler(targetSpec: DropTargetHookSpec) {
 	const targetSpecRef = React.useRef(targetSpec)
 
-	React.useEffect(() => {
+	React.useEffect(function updateDropTargetSpec() {
 		targetSpecRef.current = targetSpec
 	})
 
@@ -23,18 +19,18 @@ function useDropTargetHandler(targetSpec: DropTargetSpec<void>) {
 		() => ({
 			canDrop() {
 				const { canDrop } = targetSpecRef.current
-				return canDrop != null ? (canDrop as NoArgs<any>)() : true
+				return canDrop ? canDrop() : true
 			},
 			hover() {
 				const { hover } = targetSpecRef.current
-				if (hover != null) {
-					;(hover as NoArgs<void>)()
+				if (hover) {
+					hover()
 				}
 			},
 			drop() {
 				const { drop } = targetSpecRef.current
-				if (drop != null) {
-					;(drop as NoArgs<void>)()
+				if (drop) {
+					drop()
 				}
 			},
 		}),
@@ -44,10 +40,16 @@ function useDropTargetHandler(targetSpec: DropTargetSpec<void>) {
 	return handler
 }
 
+/**
+ * useDropTarget Hook
+ * @param ref The drop target's ref
+ * @param type The drop target type
+ * @param targetSpec The drop target specification
+ */
 export function useDropTarget(
 	ref: Ref,
 	type: TargetType,
-	targetSpec: DropTargetSpec<void> & {
+	targetSpec: DropTargetHookSpec & {
 		dropTargetOptions?: {}
 	},
 ): DropTargetMonitor & HandlerManager {
@@ -70,15 +72,15 @@ export function useDropTarget(
 
 	React.useEffect(() => {
 		const dropTargetNode = ref.current
-		const dropTargetOptions = targetSpec.dropTargetOptions
-
-		const disconnectDropTarget = backend.connectDropTarget(
-			targetMonitor.getHandlerId(),
-			dropTargetNode,
-			dropTargetOptions,
-		)
-
-		return disconnectDropTarget
+		if (dropTargetNode) {
+			const dropTargetOptions = targetSpec.dropTargetOptions
+			const disconnectDropTarget = backend.connectDropTarget(
+				targetMonitor.getHandlerId(),
+				dropTargetNode,
+				dropTargetOptions,
+			)
+			return disconnectDropTarget
+		}
 	}, [])
 
 	return targetMonitor
